Modify history storage period of skynet-zabbix items
Default all history will be stored in zabbix for about three months.
However most of the samples will not be used for that ESMonitor just
retrieve those data periodly, and then display them on the UI web only once.

When we want to draw trends graph,it's very meaningful for extending longer
history period.But most for item history data is only used once and then discard.
I hava no experice for that how long the history period is the best choice.
Accoring to default setting of zabbix-trigger type item. So i chose the value "7"

Change-Id: Ib13f150d8065c8a038079329d660ed8577a8b040